Fix nps test case
